:lock: Avoid masking return values
In the original code, the return value of `C45_Token` is ignored, and `export` will instead always return true. This may prevent conditionals, `set -e` and traps from working correctly.

When first marked for export and assigned separately, the return value of the assignment will be that of `C45_Token`. This avoids the problem.